refactor(staking): move EVM staking types to @ledgerhq/types-live
