+++ libs/coin-modules/coin-evm/src/staking/detectOperationType.test.ts
@@ -1,5 +1,5 @@
 import { ethers } from "ethers";
-import { StakingOperation } from "./../types/staking";
+import { StakingOperation } from "../types/staking";
 import { detectEvmStakingOperationType, isStakingOperation } from "./detectOperationType";
 
 describe("isStakingOperation", () => {
