+++ libs/coin-modules/coin-evm/src/staking/redelegations.ts
@@ -3,8 +3,8 @@ import network from "@ledgerhq/live-network";
 import { log } from "@ledgerhq/logs";
 import { BigNumber } from "bignumber.js";
 import type { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
-import type { Operation } from "@ledgerhq/types-live";
-import type { RedelegationStrategy, StakingRedelegation } from "../types/staking";
+import type { Operation, StakingRedelegation } from "@ledgerhq/types-live";
+import type { RedelegationStrategy } from "../types/staking";
 import { STAKING_CONTRACTS } from "./contracts";
 import { getStakingABI } from "./abis";
 import { getCoinConfig } from "../config";
