+++ libs/ledger-live-common/src/families/evm/staking/react.test.ts
@@ -4,11 +4,7 @@
 import { renderHook, waitFor } from "@testing-library/react";
 import BigNumber from "bignumber.js";
 import type { Unit } from "@ledgerhq/types-cryptoassets";
-import type {
-  StakingValidatorItem,
-  StakingAccount,
-  StakingDelegation,
-} from "@ledgerhq/coin-evm/types/index";
+import type { StakingValidatorItem, StakingAccount, StakingDelegation } from "@ledgerhq/types-live";
 import * as stakingIndex from "@ledgerhq/coin-evm/staking/index";
 import * as accountModule from "../../../account";
 import {
